+++ scripts/dzVents/documentation/history.md
@@ -7,6 +7,7 @@
  - Added adapter for the new Temperature+Barometer device.
  - Added support for groups and scenes change events. Use "on = { scenes = { 'myScene1', 'myScene2' }, groups = {'myGroup1'} }"
  - Added method backupDatabase to the domoticz object. Now you can create your own (timer) scripts to create backups.
+ - Added domoticz.startTime giving you the time at which the Domoticz service was started.
 
 [2.2.0]